+++ src/OpenHosta/core/config.py
@@ -1,26 +1,26 @@
 
 import sys
 
-from ..narrators import OneTurnConversationNarrator
-from ..models import DialogueModel, OpenAICompatibleModel
+from ..pipelines import OneTurnConversationPipeline
+from ..models import Model, OpenAICompatibleModel
 
-_DefaultNarrator = OneTurnConversationNarrator(
+_DefaultPipeline = OneTurnConversationPipeline(
     default_model=OpenAICompatibleModel(
         model_name="gpt-4o", 
         base_url="https://api.openai.com/v1/chat/completions")
 )
      
 def set_default_model(new):
-    if isinstance(new, DialogueModel):
-        _DefaultNarrator.model = new
+    if isinstance(new, Model):
+        _DefaultPipeline.model = new
     else:
         sys.stderr.write("[CONFIG_ERROR] Invalid model instance.\n")
 
 def set_default_apiKey(api_key:str=None):
     if api_key is not None or isinstance(api_key, str):
-        _DefaultNarrator.model.api_key = api_key
+        _DefaultPipeline.model.api_key = api_key
     else:
         sys.stderr.write("[CONFIG_ERROR] Invalid API key.")    
 
 def set_default_api_parameters(api_parameters:dict):
-    _DefaultNarrator.model.api_parameters |= api_parameters
+    _DefaultPipeline.model.api_parameters |= api_parameters